django-requirements.txt
django-admin-bootstrapped==0.4.3
django-debug-toolbar==0.9.4
django-discover-runner==1.0
django-endless-pagination==2.0
django-pagination==1.0.7
django-tinymce==1.5.1
django-twitter-bootstrap==3.0.0-rc2
django-uni-form==0.9.0
django-registration
